+++ fuzzer/api/src/commonMain/kotlin/com/github/jcornaz/kwik/fuzzer/api/Arbitrary.kt
@@ -1,0 +1,8 @@
+package com.github.jcornaz.kwik.fuzzer.api
+
+/**
+ * This is an empty object for declaring [Fuzzer] factories as extension function on [Arbitrary].
+ *
+ * So that one can use [Arbitrary] to easily find any [Fuzzer] factory, regardless on the module it comes from.
+ */
+object Arbitrary