+++ fuzzer/api/src/commonTest/kotlin/com/github/jcornaz/kwik/simplifier/api/SimplificationTest.kt
@@ -1,8 +1,9 @@
-package com.github.jcornaz.kwik.evaluator
+package com.github.jcornaz.kwik.simplifier.api
 
 import com.github.jcornaz.kwik.simplifier.api.ExperimentalKwikFuzzer
 import com.github.jcornaz.kwik.simplifier.api.Simplifier
 import com.github.jcornaz.kwik.simplifier.api.dontSimplify
+import com.github.jcornaz.kwik.simplifier.api.findSimplestFalsification
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -14,7 +15,7 @@ class SimplificationTest {
     fun returnsInitialValueIfThereIsNoSimplerValue() {
         repeat(100) {
             val initialValue = Random.nextInt()
-            assertEquals(initialValue, dontSimplify<Int>().simplify(initialValue) { Random.nextBoolean() })
+            assertEquals(initialValue, dontSimplify<Int>().findSimplestFalsification(initialValue) { Random.nextBoolean() })
         }
     }
 
@@ -26,7 +27,7 @@ class SimplificationTest {
 
         repeat(100) {
             val initialValue = Random.nextInt()
-            assertEquals(initialValue, simplifier.simplify(initialValue) { true })
+            assertEquals(initialValue, simplifier.findSimplestFalsification(initialValue) { true })
         }
     }
 
@@ -39,7 +40,7 @@ class SimplificationTest {
 
         repeat(100) {
             val initialValue = Random.nextInt(1, 100)
-            assertEquals(0, simplifier.simplify(initialValue) { false })
+            assertEquals(0, simplifier.findSimplestFalsification(initialValue) { false })
         }
     }
 
@@ -52,7 +53,7 @@ class SimplificationTest {
 
         repeat(100) {
             val initialValue = Random.nextInt()
-            assertEquals(null, simplifier.simplify(initialValue) { false })
+            assertEquals(null, simplifier.findSimplestFalsification(initialValue) { false })
         }
     }
 
@@ -68,7 +69,7 @@ class SimplificationTest {
 
         repeat(1000) {
             val initialValue = Random.nextInt(100, 200)
-            val result = simplifier.simplify(initialValue) { it < 42 }
+            val result = simplifier.findSimplestFalsification(initialValue) { it < 42 }
             assertEquals(42, result)
         }
     }